+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - History point storage and query call
 - History canister integration
 
+### Changed
+
+- History event implementation
+
 ### Fixed
 
 - Websocket unread count