+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [0.2.0]
 
 ### Added
 
@@ -278,7 +278,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Missing `notification_id` on migration models
 
-[Unreleased]: https://github.com/Catalyze-Software/proxy/compare/0.1.8...HEAD
+[Unreleased]: https://github.com/Catalyze-Software/proxy/compare/0.1.9...HEAD
+[0.2.0]: https://github.com/Catalyze-Software/proxy/compare/0.1.9...0.2.0
 [0.1.9]: https://github.com/Catalyze-Software/proxy/compare/0.1.8...0.1.9
 [0.1.8]: https://github.com/Catalyze-Software/proxy/compare/0.1.7...0.1.8
 [0.1.7]: https://github.com/Catalyze-Software/proxy/compare/0.1.6...0.1.7